requirements.txt
streamlit
python-dotenv
python-multipart
playwright
beautifulsoup4
openai
openpyxl